Extract config builder to separate module
